Cast string size to int to suppress compiler warning

+++ src/ui/window.cpp
@@ -71,7 +71,7 @@ void window::set_title(const std::wstring title)
 std::wstring window::get_title() const
 {
 	std::wstring text(GetWindowTextLengthW(handle) + 1, 0);
-	GetWindowTextW(handle, &text[0], text.size());
+	GetWindowTextW(handle, &text[0], (int)text.size());
 	text.resize(text.size() - 1);
 	return text;
 }
